WalletSigner interface implementation POC (#8)
